[redux] move some redux utils to shared file (#1164)
* move some redux utils to shared file so they can be used in the new export view too

* enhancer is a func now
