+++ caravel/assets/javascripts/SqlLab/reducers.js
@@ -1,6 +1,8 @@
 import shortid from 'shortid';
 import * as actions from './actions';
 import { now } from '../modules/dates';
+import { addToObject, alterInObject, alterInArr, removeFromArr, getFromArr, addToArr }
+  from '../reduxUtils.js';
 
 const defaultQueryEditor = {
   id: shortid.generate(),
@@ -22,68 +24,6 @@ export const initialState = {
   workspaceQueries: [],
   queriesLastUpdate: 0,
 };
-
-function addToObject(state, arrKey, obj) {
-  const newObject = Object.assign({}, state[arrKey]);
-  const copiedObject = Object.assign({}, obj);
-
-  if (!copiedObject.id) {
-    copiedObject.id = shortid.generate();
-  }
-  newObject[copiedObject.id] = copiedObject;
-  return Object.assign({}, state, { [arrKey]: newObject });
-}
-
-function alterInObject(state, arrKey, obj, alterations) {
-  const newObject = Object.assign({}, state[arrKey]);
-  newObject[obj.id] = Object.assign({}, newObject[obj.id], alterations);
-  return Object.assign({}, state, { [arrKey]: newObject });
-}
-
-function alterInArr(state, arrKey, obj, alterations) {
-  // Finds an item in an array in the state and replaces it with a
-  // new object with an altered property
-  const idKey = 'id';
-  const newArr = [];
-  state[arrKey].forEach((arrItem) => {
-    if (obj[idKey] === arrItem[idKey]) {
-      newArr.push(Object.assign({}, arrItem, alterations));
-    } else {
-      newArr.push(arrItem);
-    }
-  });
-  return Object.assign({}, state, { [arrKey]: newArr });
-}
-
-function removeFromArr(state, arrKey, obj, idKey = 'id') {
-  const newArr = [];
-  state[arrKey].forEach((arrItem) => {
-    if (!(obj[idKey] === arrItem[idKey])) {
-      newArr.push(arrItem);
-    }
-  });
-  return Object.assign({}, state, { [arrKey]: newArr });
-}
-
-function getFromArr(arr, id) {
-  let obj;
-  arr.forEach((o) => {
-    if (o.id === id) {
-      obj = o;
-    }
-  });
-  return obj;
-}
-
-function addToArr(state, arrKey, obj) {
-  const newObj = Object.assign({}, obj);
-  if (!newObj.id) {
-    newObj.id = shortid.generate();
-  }
-  const newState = {};
-  newState[arrKey] = [...state[arrKey], newObj];
-  return Object.assign({}, state, newState);
-}
 
 export const sqlLabReducer = function (state, action) {
   const actionHandlers = {
